requirements.txt
PyQt6>=6.4.0
cryptography>=41.0.0
requests>=2.31.0
scapy>=2.5.0
python-nmap>=0.7.1
passlib>=1.7.4
netifaces>=0.11.0
pyOpenSSL>=23.0.0
dnspython>=2.3.0
python-whois>=0.8.0 